+++ test/integration/expected_out_concat/mixed_format.py
@@ -6,4 +6,4 @@ This module will be transformed... into something far greater.
 x = 42
 thing = 1
 otherThing = '2'
-print(f"blah{thing}blah{otherThing}is {x:d}")
+print(f"blah{thing}blah{otherThing}is {x:f}")