+++ test/integration/expected_out_single_line/percent_numerics.py
@@ -1,2 +1,2 @@
 a, b, c, d, e = tuple(range(5))
-print(f'{a:d} {b:f} {c:e} {d:g} {e}')
+print(f'{a:f} {b:f} {c:e} {d:g} {e}')