+++ test/integration/expected_out_single_line/simple_percent.py
@@ -1,2 +1,2 @@
 var = 5
-a = f"Test: {var:d}"
+a = f"Test: {var}"