+++ test/integration/samples_in/double_conv.py
@@ -1,1 +1,1 @@
-print("{}".format(Bar + 1), '%d' % var, "{s}".format(s=foo))
+print("{}".format(Bar + 1), '%f' % var, "{s}".format(s=foo))