+++ test/integration/samples_in/percent_dict.py
@@ -1,3 +1,3 @@
 a = 2
 b = "wuga"
-print('%(num)d %(string)s' % {"num": a, "string": b})
+print('%(num)f %(string)s' % {"num": a, "string": b})