+++ test/integration/samples_in/percent_numerics.py
@@ -1,2 +1,2 @@
 a, b, c, d, e = tuple(range(5))
-print('%d %f %e %g %s' % (a,b,c,d,e))
+print('%f %f %e %g %s' % (a,b,c,d,e))