+++ test/integration/samples_in/simple_percent_comment.py
@@ -1,2 +1,2 @@
 var = 5
-a = "Test: %d" % var # cool formatting!
+a = "Test: %s" % var # cool formatting!