+++ test/integration/samples_in_concat/mixed_format.py
@@ -6,4 +6,4 @@ This module will be transformed... into something far greater.
 x = 42
 thing = 1
 otherThing = '2'
-print('blah{}'.format(thing) + 'blah' + otherThing + "is %d" % x)
+print('blah{}'.format(thing) + 'blah' + otherThing + "is %f" % x)